maybe performance improvement but fixed

+++ src/MinecraftDataStructures.jl
@@ -52,7 +52,7 @@ end
 
 function PooledArray(compressed::CompressedPalettedContainer, min_bits::Int64, len::Int)
   wordsize = max(min_bits, ceil(Int, log2(length(compressed.palette))))
-  data = ones(UInt64, len)
+  data = ones(UInt32, len)
   shift = 0
   j = 1
   mask = 2^wordsize - 1
@@ -70,7 +70,7 @@ function PooledArray(compressed::CompressedPalettedContainer, min_bits::Int64, l
   end
   # TODO: find a way to directly set the PooledArray
   # return PooledArray(getindex(compressed.palette, data))
-  return PooledArray(PooledArrays.RefArray(data), Dict(b => UInt64(i) for (i, b) in enumerate(compressed.palette)), compressed.palette, Threads.Atomic{Int64}(1))
+  return PooledArray(PooledArrays.RefArray(data), Dict(b => UInt32(i) for (i, b) in enumerate(compressed.palette)), compressed.palette, Threads.Atomic{Int64}(1))
 end
 
 end
